pyproject.toml
[project]
name = "nim65s-talks-index"
version = "1.0.0"
description = "Generate an index page for my talks"
readme = "README.md"
authors = [
    { name = "Guilhem Saurel", email = "guilhem.saurel@laas.fr" }
]
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
]

[project.scripts]
nim65s-talks-index = "nim65s_talks_index:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
